clean up; temps => float

+++ 3way_controller/components/lib/heater_temp.c
@@ -1,3 +1,12 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include "esp_log.h"
+#include "esp_err.h"
+#include "freertos/FreeRTOS.h"
+#include "driver/temp_sensor.h"
+#include "libconfig.h"
+#include "libdecls.h"
+
 // Get the temperature of the heater.
 // Right now we are going to try using the onboard thermometer of the esp32c chip itself.
 // This will give us a temperature reading "near" the heater.
@@ -8,16 +17,6 @@
 // See https://docs.espressif.com/projects/esp-idf/en/latest/esp32c3/api-reference/peripherals/temp_sensor.html
 // And {$IDF_SRC}/examples/peripherals/temp_sensor
 
-#include <stdio.h>
-#include <stdlib.h>
-#include "esp_log.h"
-#include "esp_err.h"
-#include "freertos/FreeRTOS.h"
-#include "driver/temp_sensor.h"
-
-#include "libconfig.h"
-#include "libdecls.h"
-
 static char *TAG = "heater temp";
 
 void heater_temp_reader_init() {
@@ -26,7 +25,7 @@ void heater_temp_reader_init() {
     ESP_ERROR_CHECK(temp_sensor_start());
 }
 
-int current_heater_temperature() {
+float current_heater_temperature() {
     float val;
     int err;
     err = temp_sensor_read_celsius(&val);
@@ -37,6 +36,6 @@ int current_heater_temperature() {
     }
     else {
         ESP_LOGI(TAG,"Heater temperature is %f", val);
-        return (int)(val);
+        return val;
     }
 }
