+++ 3way_controller/components/lib/include/libdecls.h
@@ -1,7 +1,7 @@
 
 // Temperature listener
 void temperature_listener_task();
-int current_ambient_temperature(); // in Celsius
+float current_ambient_temperature(); // in Celsius
 
 // Power controller
 void set_temperature_schedule( int temps[] );
@@ -9,4 +9,4 @@ void power_controller_start();
 
 // Heater temperature
 void heater_temp_reader_init();
-int current_heater_temperature();
+float current_heater_temperature();
